+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/dimazub/Documents/Projects/pokemon-game/src/index.js":"1","/Users/dimazub/Documents/Projects/pokemon-game/src/App.js":"2","/Users/dimazub/Documents/Projects/pokemon-game/src/Components/Layout/index.js":"3","/Users/dimazub/Documents/Projects/pokemon-game/src/Components/Header/index.js":"4","/Users/dimazub/Documents/Projects/pokemon-game/src/Components/Footer/index.js":"5","/Users/dimazub/Documents/Projects/pokemon-game/src/Components/PokemonCard/index.js":"6"},{"size":220,"mtime":1611759445161,"results":"7","hashOfConfig":"8"},{"size":4648,"mtime":1612039727823,"results":"9","hashOfConfig":"8"},{"size":812,"mtime":1612035894457,"results":"10","hashOfConfig":"8"},{"size":349,"mtime":1612036010808,"results":"11","hashOfConfig":"8"},{"size":290,"mtime":1612036020902,"results":"12","hashOfConfig":"8"},{"size":2070,"mtime":1612040490651,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"rmij52",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/dimazub/Documents/Projects/pokemon-game/src/index.js",[],"/Users/dimazub/Documents/Projects/pokemon-game/src/App.js",[],"/Users/dimazub/Documents/Projects/pokemon-game/src/Components/Layout/index.js",[],"/Users/dimazub/Documents/Projects/pokemon-game/src/Components/Header/index.js",[],"/Users/dimazub/Documents/Projects/pokemon-game/src/Components/Footer/index.js",[],"/Users/dimazub/Documents/Projects/pokemon-game/src/Components/PokemonCard/index.js",[]]
+[{"/Users/dimazub/Documents/Projects/pokemon-game/src/index.js":"1","/Users/dimazub/Documents/Projects/pokemon-game/src/App.js":"2","/Users/dimazub/Documents/Projects/pokemon-game/src/Components/Layout/index.js":"3","/Users/dimazub/Documents/Projects/pokemon-game/src/Components/Header/index.js":"4","/Users/dimazub/Documents/Projects/pokemon-game/src/Components/Footer/index.js":"5","/Users/dimazub/Documents/Projects/pokemon-game/src/Components/PokemonCard/index.js":"6","/Users/dimazub/Documents/Projects/pokemon-game/src/routes/Home/index.js":"7","/Users/dimazub/Documents/Projects/pokemon-game/src/routes/Game/index.js":"8","/Users/dimazub/Documents/Projects/pokemon-game/src/Components/MenuHeader/index.js":"9","/Users/dimazub/Documents/Projects/pokemon-game/src/Components/Menu/index.js":"10","/Users/dimazub/Documents/Projects/pokemon-game/src/Components/NavBar/index.js":"11"},{"size":220,"mtime":1611759445161,"results":"12","hashOfConfig":"13"},{"size":476,"mtime":1612293110424,"results":"14","hashOfConfig":"13"},{"size":833,"mtime":1612289364104,"results":"15","hashOfConfig":"13"},{"size":607,"mtime":1612294255041,"results":"16","hashOfConfig":"13"},{"size":290,"mtime":1612289378701,"results":"17","hashOfConfig":"13"},{"size":2239,"mtime":1612289352963,"results":"18","hashOfConfig":"13"},{"size":5288,"mtime":1612294280037,"results":"19","hashOfConfig":"13"},{"size":142,"mtime":1612289346021,"results":"20","hashOfConfig":"13"},{"size":248,"mtime":1612295023872,"results":"21","hashOfConfig":"13"},{"size":1033,"mtime":1612294341922,"results":"22","hashOfConfig":"13"},{"size":391,"mtime":1612295148443,"results":"23","hashOfConfig":"13"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},"13dj28l",{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"43","messages":"44","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"45","messages":"46","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/dimazub/Documents/Projects/pokemon-game/src/index.js",[],["47","48"],"/Users/dimazub/Documents/Projects/pokemon-game/src/App.js",[],"/Users/dimazub/Documents/Projects/pokemon-game/src/Components/Layout/index.js",[],"/Users/dimazub/Documents/Projects/pokemon-game/src/Components/Header/index.js",[],"/Users/dimazub/Documents/Projects/pokemon-game/src/Components/Footer/index.js",[],"/Users/dimazub/Documents/Projects/pokemon-game/src/Components/PokemonCard/index.js",[],"/Users/dimazub/Documents/Projects/pokemon-game/src/routes/Home/index.js",[],"/Users/dimazub/Documents/Projects/pokemon-game/src/routes/Game/index.js",[],"/Users/dimazub/Documents/Projects/pokemon-game/src/Components/MenuHeader/index.js",[],"/Users/dimazub/Documents/Projects/pokemon-game/src/Components/Menu/index.js",["49"],"/Users/dimazub/Documents/Projects/pokemon-game/src/Components/NavBar/index.js",["50"],{"ruleId":"51","replacedBy":"52"},{"ruleId":"53","replacedBy":"54"},{"ruleId":"55","severity":1,"message":"56","line":3,"column":8,"nodeType":"57","messageId":"58","endLine":3,"endColumn":10},{"ruleId":"59","severity":1,"message":"60","line":11,"column":13,"nodeType":"61","endLine":11,"endColumn":55},"no-native-reassign",["62"],"no-negated-in-lhs",["63"],"no-unused-vars","'cn' is defined but never used.","Identifier","unusedVar","jsx-a11y/anchor-is-valid","The href attribute is required for an anchor to be keyboard accessible. Provide a valid, navigable address as the href value. If you cannot provide an href, but still need the element to resemble a link, use a button and change it with appropriate styles. Learn more: https://github.com/evcohen/eslint-plugin-jsx-a11y/blob/master/docs/rules/anchor-is-valid.md","JSXOpeningElement","no-global-assign","no-unsafe-negation"]
