+++ src/Components/MenuHeader/index.js
@@ -1,0 +1,17 @@
+import Menu from '../Menu';
+import NavBar from '../NavBar';
+
+
+const MenuHeader = () => {
+    return(
+        <div>
+            
+            <NavBar>
+                <Menu />
+            </NavBar>
+        </div>
+
+    )
+};
+
+export default MenuHeader;
