+++ src/routes/Game/index.js
@@ -1,0 +1,9 @@
+const GamePage = () => {
+    return(
+        <div>
+            <h1>This is a game page!</h1>
+        </div>
+    )
+};
+
+export default GamePage;
